+++ demos/big-project/src/index.ts
@@ -1,0 +1,7 @@
+import postgres from "postgres";
+
+export function check() {
+  const sql = postgres();
+
+  sql<{ id: number }[]>`SELECT id FROM starship`;
+}