+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ts-safeql/test-utils
 
+## 0.0.5
+
+### Patch Changes
+
+- 7bf6f6a: update packages postgres dependency to 3.3.0
+- Updated dependencies [7bf6f6a]
+  - @ts-safeql/shared@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [3d3ca50]
+  - @ts-safeql/shared@0.0.4
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [69b874e]
+  - @ts-safeql/shared@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes